requirements.txt
streamlit
langchain
langchain_community
langchain-google-genai
exa-py
python-dotenv
requests
pydantic
python-decouple